net: openthread: rpc: restructure client code
Split OpenThread RPC client functions into source files
that correspond with the OpenThread code structure.
