+++ subsys/net/openthread/rpc/client/CMakeLists.txt
@@ -7,15 +7,18 @@
 zephyr_library()
 
 zephyr_library_sources(
-  ot_rpc_if_client.c
-  ot_rpc_instance_client.c
-  ot_rpc_cli_client.c
-  ot_rpc_coap_client.c
-  ot_rpc_ctrl_client.c
-  ot_rpc_commissioning_client.c
-  ot_rpc_diag_client.c
-  ot_rpc_message_client.c
-  ot_rpc_udp_client.c
+  ot_rpc_cli.c
+  ot_rpc_coap.c
+  ot_rpc_diag.c
+  ot_rpc_dataset.c
+  ot_rpc_if.c
+  ot_rpc_instance.c
+  ot_rpc_ip6.c
+  ot_rpc_link.c
+  ot_rpc_message.c
+  ot_rpc_netdata.c
+  ot_rpc_thread.c
+  ot_rpc_udp.c
 )
 
 zephyr_library_include_directories(
