+++ subsys/net/openthread/rpc/client/ot_rpc_coap.c
