+++ subsys/net/openthread/rpc/client/ot_rpc_diag.c
